task-2-template/requirements.txt
robotframework==5.0.1
pytest==7.1.2